+++ test/host-request_test.js
@@ -1,223 +1,289 @@
 var helper = require('./test_helper');
-var assert = require('assert');
-var http = require('http');
-var express = require('express');
-var app = express();
-var ac = require('../index');
-var request = require('request');
+var mocks = require('./mocks');
+var config = require('../lib/internal/config');
+var HostRequest = require('../lib/internal/host-request');
+var nock = require('nock');
+var should = require('should');
+var shouldHttp = require('should-http');
+var RSVP = require('rsvp');
 var moment = require('moment');
-var jwt = require('../lib/internal/jwt');
-var hostRequest = require('../lib/internal/host-request');
-var logger = require('./logger');
-var addon = {};
+var jwt = require('atlassian-jwt');
+var extend = require('extend');
 
 describe('Host Request', function () {
-    var server;
-    var httpClient;
+    var clientSettings = {
+        clientKey: 'test-client-key',
+        oauthClientId: 'oauth-client-id',
+        sharedSecret: 'shared-secret',
+        baseUrl: 'https://test.atlassian.net'
+    };
 
-    before(function (done) {
-        app.set('env', 'development');
-        app.use(express.urlencoded());
-        app.use(express.json());
-
-        // mock host
-        app.get('/confluence/plugins/servlet/oauth/consumer-info', function (req, res) {
-            res.set('Content-Type', 'application/xml');
-            res.status(200).send(helper.consumerInfo);
-        });
-
-        app.head("/confluence/rest/plugins/1.0/", function (req, res) {
-            res.setHeader("upm-token", "123");
-            res.status(200).end();
-        });
-
-        app.get("/confluence/rest/plugins/1.0/", function(req, res) {
-            res.json({plugins: []});
-        });
-
-        // Post request to UPM installer
-
-        app.post("/confluence/rest/plugins/1.0/", function (req, res) {
-            request({
-                url: helper.addonBaseUrl + '/installed',
-                query: {
-                    jwt: createJwtToken()
-                },
-                method: 'POST',
-                json: helper.installedPayload
-            });
-            res.status(200).end();
-        });
-
-        ac.store.register("teststore", function (logger, opts) {
-            return require("../lib/store/jugglingdb")(logger, opts);
-        });
-
-        addon = ac(app, {
-            config: {
-                "development": {
-                    store: {
-                        adapter: 'teststore',
-                        type: "memory"
-                    },
-                    "hosts": [
-                        helper.productBaseUrl
-                    ]
-                }
+    var createAddonConfig = function (opts) {
+        opts = extend({
+            jwt: {
+                validityInMinutes: 3
             }
-        }, logger);
-        server = http.createServer(app).listen(helper.addonPort, function () {
-            addon.register().then(done);
+        }, opts);
+        
+        return config({}, "development", {
+            "development": opts
         });
+    };
 
-        var settings = {
-            'sharedSecret': helper.installedPayload.sharedSecret,
-            'baseUrl': helper.productBaseUrl
+    var mockAddon = function (addonConfig) {
+        if (!addonConfig) {
+            addonConfig = {};
+        }
+
+        return {
+            logger: require('./logger'),
+            key: "test-addon-key",
+            config: createAddonConfig(addonConfig),
+            descriptor: {
+                scopes: ['READ', 'WRITE']
+            },
+            settings: mocks.store(clientSettings, clientSettings.clientKey)
         };
-        addon.settings.set('clientInfo', settings, helper.installedPayload.clientKey);
-        httpClient = hostRequest(addon, { 'userKey': 'admin' }, helper.installedPayload.clientKey);
-    });
+    };
 
-    after(function (done) {
-        server.close();
-        done();
-    });
+    function getHttpClient(addonConfig, context) {
+        if (arguments.length === 0) {
+            addonConfig = context = {};
+        } else if (arguments.length === 1) {
+            context = addonConfig;
+            addonConfig = {};
+        }
 
-    function createJwtToken() {
-        var jwtPayload = {
-            "sub": 'admin',
-            "iss": helper.installedPayload.clientKey,
-            "iat": moment().utc().unix(),
-            "exp": moment().utc().add('minutes', 10).unix()
-        };
+        var a = mockAddon(addonConfig);
+        return new HostRequest(a, context, clientSettings.clientKey);
+    }
 
-        return jwt.encode(jwtPayload, helper.installedPayload.sharedSecret);
+    function interceptRequest(testCallback, replyCallback, options) {
+        var opts = extend({
+            baseUrl: clientSettings.baseUrl,
+            method: 'get',
+            path: '/some/path/on/host',
+            httpClientContext: {}
+        }, options || {});
+
+        if (!opts.requestPath) {
+            opts.requestPath = opts.path;
+        }
+
+        var interceptor = nock(opts.baseUrl)
+                            [opts.method](opts.path)
+                            .reply(replyCallback);
+
+        var httpClient = getHttpClient(opts.addonConfig, opts.httpClientContext);
+
+        if (opts.httpClientWrapper) {
+            httpClient = opts.httpClientWrapper(httpClient);
+        }
+
+        httpClient[opts.method](opts.requestPath, function() {
+            interceptor.done(); // will throw assertion if endpoint is not intercepted
+            testCallback();
+        });
+    }
+
+    function interceptRequestAsUser(testCallback, replyCallback, options) {
+        var userKey = options.userKey;
+        delete options.userKey;
+        var opts = extend({}, options, {
+            httpClientWrapper: function (httpClient) {
+                return httpClient.asUser(userKey);
+            }
+        });
+        interceptRequest(testCallback, replyCallback, opts);
     }
 
     it('constructs non-null get request', function (done) {
-        httpClient.get('/some/path/on/host').then(function(request) {
-            assert.ok(request);
-            done();
-        });
+        interceptRequest(done, 200);
     });
 
-    it('get request has headers', function (done) {
-        httpClient.get('/some/path/on/host').then(function(request) {
-            assert.ok(request.headers);
-            done();
+    describe('Headers', function () {
+        it('get request has headers', function (done) {
+            interceptRequest(done, function (uri, requestBody) {
+                should.exist(this.req.headers);
+            });
         });
-    });
 
-    it('get request has Authorization header', function (done) {
-        httpClient.get('/some/path/on/host').then(function(request) {
-            assert.ok(request.headers['Authorization']);
-            done();
+        it('get request has user-agent header', function (done) {
+            interceptRequest(done, function (uri, requestBody) {
+                this.req.headers['user-agent'].should.startWith('atlassian-connect-express/');
+            });
         });
-    });
 
-    it('get request has Authorization header starting with "JWT "', function (done) {
-        httpClient.get('/some/path/on/host').then(function(request) {
-            assert.equal(request.headers['Authorization'].indexOf('JWT '), 0);
-            done();
+        it('get request has user-agent version set to package version', function (done) {
+            var aceVersion = require('../package.json').version;
+            interceptRequest(done, function (uri, requestBody) {
+                this.req.headers['user-agent'].should.startWith('atlassian-connect-express/' + aceVersion);
+            });
         });
-    });
 
-    it('get request has correct JWT subject claim', function (done) {
-        httpClient.get('/some/path/on/host').then(function(request) {
-            var jwtToken = request.headers['Authorization'].slice(4);
-            var decoded = jwt.decode(jwtToken, helper.installedPayload.clientKey, true);
-            assert.equal(decoded.sub, 'admin');
-            done();
-        });
-    });
+        it('post request preserves custom header', function (done) {
+            var interceptor = nock(clientSettings.baseUrl)
+                                .post('/some/path')
+                                .reply(function (uri, requestBody) {
+                                    this.req.headers.custom_header.should.eql('arbitrary value');
+                                });
 
-    it('get request has correct JWT qsh for encoded parameter', function (done) {
-        httpClient.get('/some/path/on/host?q=~%20text').then(function(request) {
-            var jwtToken = request.headers['Authorization'].slice(4);
-            var decoded = jwt.decode(jwtToken, helper.installedPayload.clientKey, true);
-            var expectedQsh = jwt.createQueryStringHash({
-              'method': 'GET',
-              'path'  : '/some/path/on/host',
-              'query' : { 'q' : '~ text'}
-            }, false, helper.productBaseUrl);
-            assert.equal(decoded.qsh, expectedQsh);
-            done();
-        });
-    });
-
-    it('post request has correct url', function (done) {
-        var relativeUrl = '/some/path/on/host';
-        httpClient.post(relativeUrl).then(function(request) {
-            assert.equal(request.uri.href, helper.productBaseUrl + relativeUrl);
-            done();
-        });
-    });
-
-    it('post request preserves custom header', function (done) {
-        httpClient.post({
-            'url': '/some/path',
-            'headers': {
-                'custom_header': 'arbitrary value'
-            }
-        }).then(function(request) {
-            assert.equal(request.headers['custom_header'], 'arbitrary value');
-            done();
-        });
-    });
-
-    it('post request with form sets form data', function (done) {
-        httpClient.post({
-            'url': '/some/path',
-            file: [
-                'file content', {
-                    filename: 'filename',
-                    ContentType: 'text/plain'
+            getHttpClient().post({
+                'url': '/some/path',
+                'headers': {
+                    'custom_header': 'arbitrary value'
                 }
-            ]
-        }).then(function(request) {
-            assert.deepEqual(request.file, ["file content",{"filename":"filename","ContentType":"text/plain"}]);
-            done();
+            }, function(request) {
+                interceptor.done();
+                done();
+            });
         });
     });
 
+    describe('Add-on JWT authentication', function () {
+        it('get request has Authorization header', function (done) {
+            interceptRequest(done, function (uri, requestBody) {
+                should.exist(this.req.headers.authorization);
+            });
+        });
 
-    it('post requests using multipartFormData have the right format', function (done) {
-        var someData = 'some data';
-        httpClient.post({
-            url: '/some/path',
-            multipartFormData: {
-                file: [someData, { filename:'myattachmentagain.png' }]
-            }
-        }).then(function(request) {
-            assert.ok(request._form);
-            assert.equal(request._form._valueLength, someData.length);
-            done();
+        it('bitbucket request sets sub claim as clientKey', function (done) {
+            interceptRequest(done, function (uri, requestBody) {
+                var jwtToken = this.req.headers.authorization.slice(4);
+                var clientKey = clientSettings.clientKey;
+                var decoded = jwt.decode(jwtToken, clientKey, true);
+                decoded.sub.should.eql(clientKey);
+            }, {
+                addonConfig: {
+                    product: 'bitbucket'
+                }
+            });
+        });
+
+        it('get request has Authorization header starting with "JWT "', function (done) {
+            interceptRequest(done, function (uri, requestBody) {
+                this.req.headers.authorization.should.startWith('JWT ');
+            });
+        });
+
+        it('get request has correct JWT qsh for encoded parameter', function (done) {
+            interceptRequest(done, function (uri, requestBody) {
+                var jwtToken = this.req.headers.authorization.slice(4);
+                var decoded = jwt.decode(jwtToken, clientSettings.clientKey, true);
+                var expectedQsh = jwt.createQueryStringHash({
+                  'method': 'GET',
+                  'path'  : '/some/path/on/host',
+                  'query' : { 'q' : '~ text'}
+                }, false, helper.productBaseUrl);
+                decoded.qsh.should.eql(expectedQsh);
+            }, { path: '/some/path/on/host?q=~%20text'});
+        });
+
+        it('get request for absolute url on host has Authorization header', function (done) {
+            interceptRequest(done, function (uri, requestBody) {
+                this.req.headers.authorization.should.startWith('JWT ');
+            }, { requestPath: 'https://test.atlassian.net/some/path/on/host' });
+        });
+
+        it('post request has correct url', function (done) {
+            interceptRequest(done, function (uri, requestBody) {
+                this.req.headers.authorization.should.startWith('JWT ');
+            }, { method: 'post' });
         });
     });
 
-    it('post requests using the deprecated form parameter still have the right format', function (done) {
-        var someData = 'some data';
-        httpClient.post({
-            url: '/some/path',
-            form: {
-                file: [someData, { filename:'myattachmentagain.png' }]
-            }
-        }).then(function(request) {
-            assert.ok(request._form);
-            assert.equal(request._form._valueLength, someData.length);
-            done();
+    describe('User impersonation requests', function () {
+        it('Request as user does not add JWT authorization header', function (done) {
+            var authServiceMock = mocks.oauth2.service();
+            interceptRequestAsUser(done, function (uri, requestBody) {
+                authServiceMock.done();
+                this.req.headers.authorization.should.not.startWith('JWT');
+            }, { userKey: 'sruiz' });
+        });
+
+        it('Request as user adds a Bearer authorization header', function (done) {
+            var authServiceMock = mocks.oauth2.service();
+            interceptRequestAsUser(done, function (uri, requestBody) {
+                authServiceMock.done();
+                this.req.headers.authorization.should.startWith('Bearer');
+            }, { userKey: 'sruiz' });
         });
     });
 
-    it('post requests using urlEncodedFormData have the right format', function (done) {
-        httpClient.post({
-            url: '/some/path',
-            urlEncodedFormData: {
-                param1: 'value1'
-            }
-        }).then(function(request) {
-            assert.equal(request.body.toString(), 'param1=value1');
-            done();
+    describe('Form requests', function () {
+        it('post request with form sets form data', function (done) {
+            var interceptor = nock(clientSettings.baseUrl)
+                        .post('/some/path')
+                        .reply(200);
+
+
+            getHttpClient().post({
+                'url': '/some/path',
+                file: [
+                    'file content', {
+                        filename: 'filename',
+                        ContentType: 'text/plain'
+                    }
+                ]
+            }).then(function(request) {
+                request.file.should.eql(["file content", {"filename":"filename","ContentType":"text/plain"}]);
+                done();
+            });
+        });
+
+
+        it('post requests using multipartFormData have the right format', function (done) {
+            var interceptor = nock(clientSettings.baseUrl)
+                        .post('/some/path')
+                        .reply(200);
+
+            var someData = 'some data';
+            getHttpClient().post({
+                url: '/some/path',
+                multipartFormData: {
+                    file: [someData, { filename:'myattachmentagain.png' }]
+                }
+            }).then(function(request) {
+                request._form.should.be.ok();
+                request._form._valueLength.should.eql(someData.length);
+                done();
+            });
+        });
+
+        it('post requests using the deprecated form parameter still have the right format', function (done) {
+            var interceptor = nock(clientSettings.baseUrl)
+                        .post('/some/path')
+                        .reply(200);
+
+            var someData = 'some data';
+            getHttpClient().post({
+                url: '/some/path',
+                form: {
+                    file: [someData, { filename:'myattachmentagain.png' }]
+                }
+            }).then(function(request) {
+                request._form.should.be.ok();
+                request._form._valueLength.should.eql(someData.length);
+                done();
+            }, function (err) {
+                console.log(err);
+            });
+        });
+
+        it('post requests using urlEncodedFormData have the right format', function (done) {
+            var interceptor = nock(clientSettings.baseUrl)
+                        .post('/some/path')
+                        .reply(200);
+
+            getHttpClient().post({
+                url: '/some/path',
+                urlEncodedFormData: {
+                    param1: 'value1'
+                }
+            }).then(function(request) {
+                request.body.toString().should.eql('param1=value1');
+                done();
+            });
         });
     });
 });